# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Magpie Admin API",
    "version": "unversioned",
    "description": "The authenticated administration surface exposed only by a GUI-capable `magpie serve` when MAGPIE_ADMIN_ADDR and MAGPIE_ADMIN_TOKEN are set. It mirrors Magpie's existing desktop handler. Desktop actions are client-owned in server mode."
  },
  "servers": [{ "url": "/" }],
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
    "/openapi.json": {
      "get": {
        "operationId": "getOpenAPI",
        "summary": "Get this OpenAPI document",
        "tags": ["State"],
        "responses": {
          "200": { "description": "OpenAPI 3.1 JSON", "content": { "application/json": { "schema": { "type": "object" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/api/open": {
      "post": {
        "operationId": "openURL",
        "summary": "Open a URL through the desktop host",
        "description": "No-op in server mode. The calling application should open the URL locally.",
        "tags": ["Host"],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/URLRequest" } } } },
        "responses": { "204": { "description": "Accepted; no browser is opened in server mode" }, "401": { "$ref": "#/components/responses/Unauthorized" } }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "description": "The exact MAGPIE_ADMIN_TOKEN value." }
    },
    "responses": {
...
      "Unauthorized": { "description": "Missing or invalid bearer token", "headers": { "WWW-Authenticate": { "schema": { "type": "string", "const": "Bearer" } } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
    },
    "schemas": {
      "Error": { "type": "object", "required": ["error"], "properties": { "error": { "type": "string" } } },
...
      "URLRequest": { "type": "object", "required": ["url"], "properties": { "url": { "type": "string", "format": "uri" } } },
...
    }
  }
}